requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.7
daphne==4.0.0
Django==5.0.1
django-apscheduler==0.6.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.3.1
django-import-export==4.1.1
django-redis==5.4.0
django-timezone-field==6.1.0
django-vi-address==0.1.6
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
Faker==22.4.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
kombu==5.3.7
msgpack==1.0.7
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
django-vi-address==0.1.6
payos=0.1.7